Moved ToyPad emu logic into ToyPadEmu.c
Updated references throughout the codebase to use the new ToyPadEmu API, removed duplicate logic, and improved encapsulation a bit.
