+++ .gitignore
@@ -5,5 +5,3 @@ dist/*
 .editorconfig
 .env
 .ufbt
-.idea
-.DS_Store
