+++ usb/save_toypad.c
@@ -7,8 +7,6 @@
 #include "../minifigures.h"
 #include "../debug.h"
 
-#define TAG "LDToyPad"
-
 #define FILEPATH_SIZE          128
 #define FILE_NAME_LEN_MAX      256
 #define TOKEN_FILE_EXTENSION   ".toy"
@@ -307,6 +305,7 @@ void mkdir_toypads(void) {
     furi_record_close(RECORD_STORAGE);
 }
 
+// Toypad DTO (Data Transfer Object)
 typedef struct {
     Token tokens[MAX_TOKENS];
     BoxInfo boxes[NUM_BOXES];
