+++ usb/usb_toypad.h
@@ -4,6 +4,7 @@
 #include <pthread.h>
 
 #include "frame.h"
+#include "ToypadEmu.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -12,25 +13,6 @@ extern "C" {
 #define HID_EP_SZ  0x20 // 32 bytes packet size
 #define HID_EP_IN  0x81
 #define HID_EP_OUT 0x01
-
-#define MAX_TOKENS 7
-
-#define FRAME_TYPE_RESPONSE 0x55
-#define FRAME_TYPE_REQUEST  0x56
-
-typedef struct {
-    unsigned char index;
-    unsigned int id;
-    unsigned int pad;
-    unsigned char uid[7];
-    uint8_t token[180];
-    char name[16];
-} Token;
-
-typedef struct {
-    Token* tokens[MAX_TOKENS];
-    uint8_t tea_key[16];
-} ToyPadEmu;
 
 enum ConnectedStatus {
     ConnectedStatusDisconnected = 0,
@@ -45,18 +27,8 @@ extern FuriHalUsbInterface usb_hid_ldtoypad;
 
 usbd_device* get_usb_device();
 
-bool ToyPadEmu_remove(int index);
-
-void ToyPadEmu_clear();
-
-Token* createCharacter(int id);
-
-Token* createVehicle(int id, uint32_t upgrades[2]);
-
-int get_connected_status();
-void set_connected_status(int status);
-
-void calculate_checksum(uint8_t* buf, int length, int place);
+uint8_t get_connected_status();
+void set_connected_status(enum ConnectedStatus status);
 
 #ifdef __cplusplus
 }
